requirements.txt
Flask==1.1.2
pymongo==4.1.1
pyyaml==6.0
Flask-PyMongo==2.3.0
PyJWT==1.7.1
flask-cors==3.0.10
flask-socketio==5.2.0
simple-websocket==0.5.2
eventlet==0.33.1